poprawione metody z klasami do rejestrowania procesu
