--- zarejestruj_zjazd.rb
@@ -1,10 +1,0 @@
-def cli_zarejestruj_powrot
-    puts "Podaj datę zjazdu"
-    @data_zjazdu = gets.chomp
-   
-    puts "Podaj stan licznika po powrocie z trasy"
-    @km_powrot = gets.chomp.to_f
-      
-    puts "Podaj stan paliwa po powrocie z trasy"
-    @paliwo_zjazd = gets.chomp.to_f
-   end
